refactor application runtime interface and move to `runtime.rs`
