+++ examples/snake/src/main.rs
@@ -2,9 +2,8 @@ mod game;
 use crate::game::create_game;
 use hewn::game::GameHandler;
 #[cfg(not(target_arch = "wasm32"))]
-use hewn::runtime::{initialize_terminal_io, TerminalRuntime};
-#[cfg(not(target_arch = "wasm32"))]
-use hewn::view::{ScreenDimensions, TerminalRenderer, View};
+use hewn::runtime::TerminalRuntime;
+use hewn::runtime::WindowRuntime;
 
 const SCREEN_WIDTH: u16 = 50;
 const SCREEN_HEIGHT: u16 = 50;
@@ -12,7 +11,9 @@ const SCREEN_HEIGHT: u16 = 50;
 fn main() {
     let mut game = create_game(SCREEN_WIDTH, SCREEN_HEIGHT, None);
     game.start_game();
-    hewn::render::app::run(Box::new(game)).unwrap();
+    let mut runtime = WindowRuntime::new();
+    let _ = runtime.start(&mut game);
+
     let mut game = create_game(SCREEN_WIDTH, SCREEN_HEIGHT, None);
     let mut runtime = TerminalRuntime::new(SCREEN_WIDTH, SCREEN_HEIGHT);
     runtime.start(&mut game);
