+++ src/render.rs
@@ -1,2 +1,1 @@
-pub mod app;
 pub mod render;
